install tsx, change files in scripts/ and webpack-plugins directories to be of type ts instead of js, standardise file naming convention to be kebab-case, include html files in prettier format
