+++ package.json
@@ -4,8 +4,8 @@
     "main": "index.js",
     "private": true,
     "scripts": {
-        "format": "prettier --write \"**/*.{ts,tsx,js,jsx,json,md}\"",
-        "lint": "eslint \"**/*.{ts,tsx,js,jsx,json,md}\"",
+        "format": "prettier --write \"**/*.{ts,tsx,js,jsx,json,md,html}\"",
+        "lint": "eslint \"**/*.{ts,tsx,js,jsx,json,md,html}\"",
         "typecheck": "tsc --noEmit",
         "test": "jest",
         "dev:chromium": "webpack --watch --mode development --env browser=chromium --config engines/chromium/webpack.config.js",
@@ -53,6 +53,7 @@
         "ts-jest": "^29.2.5",
         "ts-loader": "^9.5.2",
         "ts-node": "^10.9.2",
+        "tsx": "^4.19.2",
         "typescript": "^5.7.3",
         "typescript-eslint": "^8.22.0",
         "webpack": "^5.97.1",
