+++ public/options.html
@@ -1,12 +1,12 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta content="width=device-width, initial-scale=1.0" name="viewport">
-    <title>ClintonCAT options</title>
-</head>
-<body>
-<div id="root"></div>
-<script src="options.js"></script>
-</body>
+    <head>
+        <meta charset="UTF-8" />
+        <meta content="width=device-width, initial-scale=1.0" name="viewport" />
+        <title>ClintonCAT options</title>
+    </head>
+    <body>
+        <div id="root"></div>
+        <script src="options.js"></script>
+    </body>
 </html>
