+++ src/browser-extension.test.ts
