+++ src/contentscanners/amazon/amazon-us.ts
