+++ src/utils/escape-regex.ts
