--- tests/www/3products_1cat.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Product Showcase</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-<h1>Our Products</h1>
-<div class="container">
-    <div class="product" id="product1">
-        <img src="images/laptop.png" alt="Product 1">
-        <h2>Laptop</h2>
-        <p>This is a great product that you will love.</p>
-    </div>
-    <div class="product" id="product2">
-        <img src="images/gamepad.png" alt="Product 2">
-        <h2>Gamepad</h2>
-        <p>This product is amazing and has great features.</p>
-    </div>
-    <div class="product" id="product3">
-        <img src="images/brickphone.png" alt="Product 3">
-        <h2>LG G4 Phone</h2>
-        <p>Get this product today and enjoy its benefits.</p>
-    </div>
-</div>
-</body>
-</html>
