+++ tsconfig.json
@@ -4,7 +4,7 @@
         "target": "ES6",
         "lib": ["DOM", "ES6"],
         "outDir": "dist",
-        "rootDir": "src",
+        "rootDir": ".",
         "strict": true,
         "moduleResolution": "node",
         "esModuleInterop": true,
