+++ webpack-plugins/touch-manifest-plugin.js
